ほぼ動く

+++ Src/Actor/IAgent.h
@@ -9,6 +9,6 @@ public:
 	virtual const Vector2& get_position() const = 0;
 	virtual void move_towards(const Vector2& target, float speed) = 0;
 
-	virtual const int get_health() const = 0;
+	virtual const float get_health() const = 0;
 	virtual void attack() = 0;
 };
